Review specialty

+++ server/server/DTO/ServiceDTO.cs
@@ -15,7 +15,7 @@ namespace server.DTO
              public int ServiceId { get; set; }
              public string? ServiceName { get; set; }
              public string? Description { get; set; }
-             public decimal? Price { get; set; }
+             public float? Price { get; set; }
          }
      }
  }
